# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Docprod Telegram Mini App API",
    "description": "Local Telegram Mini App backend. No paid providers.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Get Usage",
        "operationId": "getUsage",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UsageTxnView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "stars": {
            "type": "integer",
            "title": "Stars"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "stars",
          "simulated",
          "created_at"
        ],
        "title": "UsageTxnView"
      },
      "UsageView": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/UsageTxnView"
            },
            "type": "array",
            "title": "Transactions"
          },
          "stars_debited": {
            "type": "integer",
            "title": "Stars Debited"
          },
          "stars_purchased": {
            "type": "integer",
            "title": "Stars Purchased"
          }
        },
        "type": "object",
        "required": [
          "transactions",
          "stars_debited",
          "stars_purchased"
        ],
        "title": "UsageView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}